+++ NEWS.md
@@ -1,4 +1,8 @@
 # purrr 0.1.0.9000
 
+* Experimental functions `maybe()`, `outputs()`, and `safe()` for working
+  with functions that have side-effecty outputs (e.g. printed output,
+  messages, warnings, and errors) (#120).
+
 * Implement `set_names()`: a camel case alternative to `setNames()` with
   stricter equality checking (#119).